Add handler event type dispatch test
